==> notebooks/0-fetch-data.ipynb <==
!pip install wget

import os
import json
import wget
import zipfile

with open("./config.json", "r") as fp:
    config = json.load(fp)

DATA_DIR_PATH = config["data_dir_path"]

os.makedirs(DATA_DIR_PATH, exist_ok=True)

# Fetch SNLI datasets

SNLI_URL = "https://nlp.stanford.edu/projects/snli/snli_1.0.zip"
SNLI_ZIP_PATH = os.path.join(DATA_DIR_PATH, "snli_1.0.zip")
SNLI_DIR_PATH = os.path.join(DATA_DIR_PATH, "snli_1.0")
print(SNLI_URL, SNLI_ZIP_PATH, SNLI_DIR_PATH)

if os.path.isfile(SNLI_ZIP_PATH) == False and os.path.isdir(SNLI_DIR_PATH) == False:
    wget.download(SNLI_URL, SNLI_ZIP_PATH)

if os.path.isdir(SNLI_DIR_PATH) == False:
    with zipfile.ZipFile(SNLI_ZIP_PATH, 'r') as zf:
        for name in zf.namelist():
            if "__MACOSX" in name or ".DS_Store" in name or "Icon" in name:
                continue
            zf.extract(name, DATA_DIR_PATH)

# Fetch Glove 840B 300D data

GLOVE_URL = "http://www-nlp.stanford.edu/data/glove.840B.300d.zip"
GLOVE_ZIP_PATH = os.path.join(DATA_DIR_PATH, "glove.840B.300d.zip")
GLOVE_FILE_PATH = os.path.join(DATA_DIR_PATH, "glove.840B.300d.txt")
print(GLOVE_URL, GLOVE_ZIP_PATH, GLOVE_FILE_PATH)

if os.path.isfile(GLOVE_ZIP_PATH) == False and os.path.isfile(GLOVE_FILE_PATH) == False:
    wget.download(GLOVE_URL, GLOVE_ZIP_PATH)

if os.path.isfile(GLOVE_FILE_PATH) == False:
    with zipfile.ZipFile(GLOVE_ZIP_PATH, 'r') as zf:
        for name in zf.namelist():
            zf.extract(name, DATA_DIR_PATH)
