==> notebooks/4-learn-esim-model.ipynb <==
# Fetch ESIM from github

!git clone https://github.com/coetaur0/ESIM ./ESIM

!pip install torch==1.4.0

# Install ESIM

!(cd ./ESIM && pip install --upgrade .)

import torch

torch.backends.cudnn.enabled = False

torch.cuda.is_available()

# Train with SNLI

!(cd ./ESIM/scripts/training && python3 train_snli.py --config="../../../esim_config.json")
